requirements.txt
langchain
openai
streamlit
tiktoken
python-dotenv
pinecone-client